+++ dailyHOTEL/app/src/main/java/com/daily/dailyhotel/screen/home/stay/inbound/filter/StayFilterActivity.java
@@ -42,7 +42,7 @@ public class StayFilterActivity extends BaseActivity<StayFilterPresenter>
 
     public static Intent newInstance(Context context, String checkInDateTime, String checkOutDateTime//
         , DailyCategoryType categoryType, String viewType, StayFilter filter, StaySuggestV2 suggest//
-        , ArrayList<String> categories, Location location, double radius, String searchWord)
+        , ArrayList<String> categories, Location location, float radius, String searchWord)
     {
         Intent intent = new Intent(context, StayFilterActivity.class);
         intent.putExtra(INTENT_EXTRA_DATA_CHECK_IN_DATE_TIME, checkInDateTime);
